dcp-815 separate supplemented_entity.py
