+++ exporter/graph/crawler.py
@@ -5,11 +5,12 @@ from operator import iconcat
 from typing import List, Iterable, Optional, Callable
 
 from exporter.graph.experiment_graph import ExperimentGraph, SupplementaryFileLink, \
-    SupplementedEntity, SupplementaryFile
+    SupplementaryFile
 from exporter.graph.input import Input
 from exporter.graph.output import Output
 from exporter.graph.process_link import ProcessLink
 from exporter.graph.protocol_link import ProtocolLink
+from exporter.graph.supplemented_entity import SupplementedEntity
 from exporter.metadata.resource import MetadataResource
 from exporter.metadata.service import MetadataService
