+++ exporter/graph/experiment_graph.py
@@ -3,19 +3,8 @@ from dataclasses import dataclass
 from typing import List, Dict, Iterable, Any, Union
 
 from exporter.graph.process_link import ProcessLink
+from exporter.graph.supplemented_entity import SupplementedEntity
 from exporter.metadata.resource import MetadataResource
-
-
-@dataclass
-class SupplementedEntity:
-    entity_type: str
-    entity_id: str
-
-    def to_dict(self) -> Dict[str, str]:
-        return dict(
-            entity_type=self.entity_type,
-            entity_id=self.entity_id
-        )
 
 
 @dataclass
